Clean up Tuya Power Config
